MM-30708: Fix race in TestLdapConnect (#9)
The core issue is that we are accessing the Debug field
after the connection reader has started.

There are 2 ways to solve this:
- Either we pass the Debug field in the constructor itself.
This would change signatures of Dial, DialTLS and NewConn.
It works but it does not look very good and is anyways an API change.
- We do not start the connection in Dial/DialTLS, but rather
start it by calling the Start method explicitly.

I have chosen the latter because functionally, it makes more sense to me.
Dial should just dial, and then we can use the Start method ourselves
after we have set up the struct to our liking.

Yes, this is a breaking change in functionality. But I don't see any other
way to do it without introducing an API change in any way. And this is
only used by us, and only in 2 places in the ldap code.

Once this goes in, we will need another PR in the EE codebase to change
the functionality and import the new code.

https://mattermost.atlassian.net/browse/MM-30708